Move json outside Match History
