+++ client/src/components/MatchHistoryPage/MatchHistoryDisplay/SummonerSpellsDisplay.tsx
@@ -1,5 +1,5 @@
 import { ParticipantDto } from "../../../types/MatchV5";
-import summonerSpells from "../json/summoner-spells.json";
+import summonerSpells from "../../json/summoner-spells.json";
 
 const SummonerSpellsDisplay = ({ playerData }: { playerData: ParticipantDto }) => {
   // Find image links for summoner spells
